Remove all non-equality "Version constraint" implementation
This patch removes the idea of "version constraints" except for the
equality constraint.  This is due to the fact, that everything else
might result in impurities.

This might be reverted in the future and actual operators ("<" or ">" or
ranges...) might be implemented. Thus, we keep the "=" equality sign as
prefix for a version string, to be extensible here.

This commit also fixes (automatically, because the implementation
changed from the ground up) the issue that there was no difference
between a version string and a version constraint string.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
(cherry picked from commit 7bbcca0356795ac60bf7761819b56430e0905a3c)