Derive Serialize for all types
Because we want to write the whole Tree object into the database as
JSON, we need to be able to serialize all these types.

Some of them (which had no implementation of Deserialize) also got an
implementation of Deserialize, because we also want to be able to read
them from the database and print them to the user.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
(cherry picked from commit 3170c9ed14e22c5baa41618326a4db02389da445)