Reorder build() impl, make use of async-await
This patch re-orderes the build() implementation.
First goes everything that is not possible to run in parallel, such as
commandline option fetching and so on.

Then, we create all the futures we need to do things.
Later, we collect them as needed, so we can run everything as late as
possible.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>