Improve scheduling implementation
This patch improves the EndpointScheduler::schedule_job() implementation
(the routine that selects the endpoint)
to actually distribute tasks over endpoints instead of passing jobs to
the first endpoint until it is full.

The current issue with the implementation is, that the write lock is
held almost all the time, which results in very little time without a
write-lock held - which might cause issues down the line when other
threads cannot take the write lock to decrease the number of current
running jobs on the endpoint.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>