Fix: Do not actually write package tree into database
This is a quickfix to get to a state where we can push jobs to the
endpoints.
The "fix" is to not serialize the tree into a JSON object, because we
cannot do this with the key of a tree entry being a non-String.

We have to work around this limitation of JSON, but this is not the time
to do it.
This patch should be reverted as soon as possible, and a real fix be
implemented.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>