Temp Fix: Collect log receivers
This is another temporary fix to get everything working.
The issue was, that the receivers were immediately dropped and did not
life long enough so that the sending side could actually write to the
channels.

This keeps the receivers alive.

The real fix is to actually write the logs to some destination.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>