Add runtime-dependencies in runnable job
This patch adds the runtime dependencies into the runnable job.
These might be required to execute tests for a package inside the
container.
Because of this we also copy them to the container.

The patch also generalizes the preparing mechansim into a helper
function and makes use of some nice iterator chaining to be more
readable.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>