Degrade ProgressBars to a simple progress-bar-generation helper
This patch removes the idea of having a "MultiProgress" object from
indicativ available in the top level of butido.

It turned out that the progress bars were not updated until the
MultiProgress::join() method was called, which basically yields the
progress bars useless.

Thus, we remove the multi-progress-bar idea from top level.
Later, we might implement a multi-progress bar viewing in the
orchestrator, so we can see which job is progressed how far.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>