Fix: build_package_filter_by_dependency_name() filters can fail
This fixes an issue with the build_package_filter_by_dependency_name()
function, which generated a filters::filter::Filter<Package>.

Though, these filters can fail because the dependency string it filters
for must be parsed to be matched against.
Because the `Package` type does not store the parsed information, but
the "raw" string from the TOML, this parsing process is necessary.

This patch changes the function to return a failable filter (a filter
that returns `Result<bool>`).
These results must be handled in the main() function (actually in the
subcommand implementation), which is done in this patch as well.

Tests were adapted accordingly, which required adding a proper version
constraint string in the test Package instances.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>