Change Tree implementation to be a Vec<_>
This patch changes the implementation of the Tree type to be a
Vec<(Package, Tree)> internally.

This is needed because we need to be able to serialize the tree into a
JSON document later, which we want to store in the database.
Because a `Package` is not a `String`, which is what is required for a
key in JSON, we cannot use `BTreeMap<Package, Tree>` for the Tree
implementation.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
Fixes: d9f6c396808b5cbee18ca2f37e4eb82df215d5fd ("Fix: Do not actually write package tree into database")