Generate staging-dir directory name
This patch changes the staging-dir selection be generating a directory
name from a UUID for the staging directory.

This is required because the staging directory must be unique per
butido-build call.

A command to overwrite the staging directory is added (for when
re-starting failed builds during debugging a package build).

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>