Move log collecting to endpoint code
This patch moves the log-setup and -collecting to the endpoint
implementation, that takes care of running the job.

This is required because here we aggregate the log and here we know all
information that need to be written to the database, so it seems natural
to move the log-collecting code (and everything else related to the
database entry for the job) to this position in the code.

This implies that the "job" is written to the database _after_ it was
run.
This is due to the fact that we cannot write the job entry before we
have the logs.
We _have to_ collect the logs in memory (we can also write to a
log-dump-file, of course), but we cannot create the job object in the
database and then append the logs to it.
This might change in the future, but for now that's the way it is done.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>