Implement error reporting of failed jobs
This patch implements error reporting if a container job did not end
successfully.

It does so by adding an error type `ContainerError`, which is either an
error that describes that a container did not exit with success, or an
anyhow::Error (that describes an error from the container management
code).

The algorithm of log-aggregation is now intercepted to catch any
exit-state log items.
If there is no exit-state from the container (No line with
"#BUTIDO:STATE:..."), no error is assumed.

Here could be a warning later on.

The so aggregated state is then passed up to the orchestrator, which
then collects the errors and prints them.

If the implementation is correct (which is not tested yet, because this
is rather difficult to test), all other containers should continue
operation until they are ready, before the errors are handled.
The code responsible for this (in the Orchestrator implementation) was
adapted to not collect until the first error, but collect everything and
then check for errors.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>