Update dependency: ascii_table must be 3.0.2 or newer
This patch updates the ascii_table dependency with a requirement to be
3.0.2 or newer.

This is because ascii_table was licensed as GPLv3 before that version,
and got relicensed to MIT in 3.0.2.

Because we do not want to depend on a library that is GPLv3, we asked
the author to relicense their crate, which is what they did for us.

This patch thus ensures that we do not link GPLv3 code.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>