Add strict script interpolation
This patch adds strict script interpolation, which means that the script
interpolation will result in an error if a variable is referenced that
does not exist.
Before this patch, referencing an absent variable did result in an empty
string, possibly resulting in an error at runtime.

This feature is on by default.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>