Make sure only the relative pathes are written to database
This patch is a bit of a mess.

It makes sure that an crate::filestore::Artifact only knows its
_relative_ path to the store it is put into.  First of all, this is what
it should be like. Secondly, we only want to track relative pathes in
the database (to reduce database size, but more
importantly because it is just duplicated data).  The full path of an
Artifact can always be reconstructed from the submit id, the configured
pathes of the staging/release store and the artifact
information (relative path) tracked in the database.

This patch would not be necessary if we would introduce strong typing
for the different kinds of pathes.
This is most certainly on our todo list here.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>