Remove passing of additional env variables
This patch removes the passing around of additional environment
variables that were specified on the commandline and adds them directly
to the Job object instance upon creation.

This does not result in a netto-loss of code, but in a netto-loss of
complexity.

For this to be possible, we had to derive Clone for `JobResource`, which
we have to clone when creating the `Job` objects during the creation of
the jobsets from the `Tree` object.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>