Add deny.toml for cargo-deny
This patch adds the deny.toml configuration file for cargo-deny to

* check licenses of the dependencies
* check that dependencies are fetched from trusted sources

We do not allow copyleft licenses, except for MPL-2.0, which is
considered "safe" as a dependency because it only applies to the code
that's licensed, not code that depends on it. (IANAL)
We allow either OSI or FSF approved "free" licenses.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>