Remove TODO because we hard-code dependency versions
This patch removes the TODO entry because this is not a TODO anymore.

Because we hard-code versions in dependencies (no ranges are allowed
anymore, only exact versions), we cannot run into a situation where a
duplicated package in a dependency tree is resolvable and NOT a hard
error.

As soon as there is a duplicate, something is wrong with the
dependency specification somewhere in the tree.
This is thus a hard error and cannot be resolved, which makes the TODO
obsolete.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>