Make sources named
This patch changes the sources to be named.

This is required for nice interpolation when layering pkg.toml files.
Before this, we had to make sure that the `[[sources]]` array element
was on the right position.

For example, consider two files:

    /pkg.toml
    /package/pkg.toml

in the first:

    [[sources]]
    url = "some/thing"

in the second:

    [[sources]
    hash.type = "sha1"
    hash.value = "asdf..."

When merged:

    [[sources]
    url = "some/thing"
    hash.type = "sha1"
    hash.value = "asdf..."

but because the _position_ was responsible for merging these elements,
adding a new source file _before_ the existing on in /pkg.toml would
yield the /package/pkg.toml incomplete when merged onto /pkg.toml, thus
resulting in an error.

With named source entries, this is less likely.

The source file name includes the source name as well, of course.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>