Add example package repository
This patch adds an example package repository including a "source
repository" where sources for packages are located.

The Makefile can be used to setup an testing environment, controlled via
environment settings (which are created via shell.nix for example) and
spin up a webserver for the sources.

Butido can then be used to download and build the sources into packages.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>