Change implementation to return Ok(()) on success
This changes the source file hash verification mechanism to return an
Ok(()) on success and an Err(_) on failure.

Returning a bool seems to be the right way, but it is not because if the
verification fails, we print the error message anyways, so there is no
point in having a bool around.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>