Add feature: script linting
This patch adds script linting via a configurable script.

With this patch, the configuration can point to a script or program that
gets the packaging script on STDIN and might exit with a nonzero value
on error.

The simplest script I can think of that adds value is:

    #!/bin/bash
    shellcheck -

to run the packaging script through shellcheck.

stdout and stderr of the linting script are printed in case of non-zero
return value.

Linting can be disabled by not setting the script in the config or by
passing a commandline flag to skip linting (a warning will be printed in
this case).

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>