Rewrite container error reporting
This patch rewrites the container error reporting to be more simple.

The ContainerError type was rewritten to not wrap other errors anymore,
but be the root error itself.
It only has one variant now, because there's only one kind of error: "It
didn't work".

The reporting in the calling functions can now use anyhow::Result<_>
instead of std::result::Result because of that.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>