Add more verbose error reporting
This patch adds more verbose error reporting in case of build error.

It alters the Orchestrator::run() interface to return the Job UUID
alongside the error object.
The UUID object can then be (and is) used in the "build" subcommand
implementation to fetch information about the failed job from the
database and print it to the user.

The number of log lines is configurable.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>