Disable logging in handlebars
Because the scripts we run should not be able to access the logger
during handlbars rendering.

This is due to the fact that we need to have control over stdout/stderr
because of the progress bar printing, but also because there's
absolutely no need for the user to log during the handlebars rendering
process.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>