Fix: Add constraints that artifact entries must be unique by path and job id
This patch alters the database so that artifact entries in the db are
unique by path and job they belong to.

The `db::models::Artifact` implementation was updated to select the
newly created Artifact object by path and by job id.

The uniqueness constraint on the `path` column in the `artifacts` table
was dropped.
Uncool: We did not generate the constraint name explicitely. That's
unfortunate as we have to gamble here that the constraint name is always
guaranteed to be the same ("artifacts_path_key").

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>