Fix: Remove indentation from script snippets when building script
We actually need to make an ugly (and very bad performing) hack here, to
get the `unindent` crate to actually unindent here.

    text = format!("\n{}", text).unindent(),

the format!() call is necessary to tell `unindent::Unindent::unindent()`
that the intended indentation is actually zero.
If we do not add that `\n` here, the first line of the to-be-unindented
indented text stays unintentionally indented.

For example, unindenting the intentionally indented string:

"   intentionally indented
    intentionally indented
"

results unintentionally in the partially indented and partially
unindented text:

"    intentionally indented
intentionally indented
"

Which is not our intention of course, although it seems to be the
intention of the unindent crate, as documented.

But with the intentionally added `\n` we unindent the indentation of all
indented "intentionally indented" lines.

Signed-off-by: Christoph Prokop <christoph.prokop@atos.net>
Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Christoph Prokop <christoph.prokop@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>