Refactor: Chain iterators and collect at once
This improves the collection building be reducing the asyncness.
The stream is build first and then all futures are collected. Before
this patch, two streams were collected and then appended.

Not sure whether this improves runtime, but it certainly improves
readability.

Unfortunately, collecting the additional job resources right into the
streams is not that easy and thus the resulting collection is simply
extend()ed.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>