Pin shiplift to revision
With this patch, we pin shiplift to a revision.
That is because shiplift may break from one revision to another on master (which
just happened and which is what we fixed in
cc8d3a7912203c90a6591e0d178c80f620bd61d0 ("Fix: Add type annotation").

We expect shiplift to hit tokio 1.0 anytime now and because of that we might
want to do a manual update of shiplift and tokio rather than an imperative
`cargo update` which may (short-term) break our codebase.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>