Fix: Spawn filesystem task on a blocking tokio thread
This patch changes the (possibly heavy) File reading from disk to be executed on
a blocking tokio thread.
This way, we ensure we do not block until the `read()` operation is finished and
can continue executing async tasks on the runtime.

Because IO might be expensive (think of a 5GB file that is read to memory), this
is a good optimization.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>