Fix: Hash verification not async
Because we continuously get blocking filesystem operations when the
implementation of the verification is async, simply remove the asyncness here
now.

This does not decrease performance (yet), because the function is called
concurrently with other calls anyways.
It blocks the tokio worker thread tho, thus maximum parallelism might be =
number of cores. :-(

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>