Fix: Progress reporting
This patch fixes progress reporting. Because our progress-bar-creating helper
initializes the bar with length 1, we have to set the length here manually.

The bar has to be added to the multibar object right away, because otherwise it
will be rendered to the output which gives us an ugly dead progress bar.
If the length is set after the adding to the multibar object, this does not
happen.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>