Fix: Ensure job UUIDs are propagated through whole tree
This patch changes the propagation of results, so that the UUIDs of the jobs
producing the artifacts are propagated through the whole tree.

This issue at hand was that when having a dependency tree like this:

    C -> B -> A

The results from A were propagated to B and the results from B where propagated
to C. But, because the implementation did not do this, the results from A where
included in the results from B and the UUID from A was dropped.

This was an issue because the implementation waited for _all_ dependencies
(direct and transitive) by their job UUID.
This means that C waited on a UUID that described the Job for A, but never
received it, which caused everything to fail.

This patch changes the algorithm, to not only report the own UUID and all
artifacts of a job, but all artifacts with their UUID attached, which solves the
issue.

The root of the tree (the `Orchestrator`) simply drops the UUIDs before
returning the artifacts to its caller.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>