Rewrite package organizational structure using DAG
This patch reimplements the package orchestration functionality to rely
on a DAG rather than a tree.

        A
       / \
      B   E
     / \   \
    C   D   F

Before this change, the structure the packages were organized in for a
build was a tree.
That did work reasonable well for initial development of butido, because
this is a simple case and the implementation is rather simple, too.

But, packages and their dependencies are not always organized in a tree.
Most of the time, they are organized in a DAG:

      .-> C -,
     /        \
    D          > A
     \        /
      `-> B -´

This is a real-world example: A could be a common crypto-library that I
do not want to name here.
B and C could be libraries that use the said crypto-library and D could
be a program that use B and C.
Because said crypto-library builds rather long, building it twice and
throwing one result away is a no-go.
A DAG as organizational structure makes that issue go away entirely.

Also, we can later implement checks whether the DAG contains multiple
versions of the same library, if that is undesireable.

The change itself is rather big, frankly because it is a non-trivial
change the replace the whole data structure and its handling in the
orchestrator code.

First of all, we introduce the "daggy" library, which provides the DAG
implementation on top of the popular "petgraph" library.

The package `Tree` datastructure was replaced by a package `Dag`
datastructure. This type implements the heavy-lifting that is needed to
load a package and all its dependencies from the `Repository` object.

The `JobTree` was also reimplemented, but as `daggy::Dag` provides a
convenient `map()` function, its implementation which transforms the
package `Dag` into a job `Dag` is rather trivial.
`crate::job::Dag` then provides the convenience `iter()` function to
iterate over all elements in the DAG and providing a `JobDefinition`
object for each node.
The topology in which we traverse the DAG is not an issue, as we need to
create tasks for all `JobDefinition`s anyways, so we do not care about
traversal topology at all.

The `crate::package::Package` type got an `Hash` implementation, which
is necessary to keep track of the mappings while reading the DAG from
the repository.
The implementation does not create the edges between the nodes in the
DAG right when inserting, but afterwards.
To keep track of the `daggy::NodeIndex`es, it keeps a mapping

    Package -> NodeIndex

in a Hashmap. Thus, `Package` must implement `std::hash::Hash`

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
Tested-by: Matthias Beyer <mail@beyermatthias.de>

squash! Reimplement as DAG