Change implementation to allow multiple senders
This patch reimplements the result-propagation to allow multiple
senders.

               .-> E
              /
      .-> C -<
     /        \
    D          >-> A
     \        /
      `-> B -´

In this scenario, A needs to send its results to multiple other jobs.

This patch changes the implementation so that each JobTask has a Vec<_>
of senders, which allows sending results to multiple parents.

In case of error, the error is only propagated to one parent, because it
doesn't matter. If there is an error, we fail and abort the whole tree
anyways, so the other parents don't need to be notified of the error.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>