Change implementation to use HashMap for storing results
This changes the implementation to use a hashmap for storing the
results.
This way, we are not storing the same result twice.

      .-> C -,
     /        \
    D          >-> A
     \        /
      `-> B -´

In this scenario, D gets the result from A propagated via B and via C.
Because of this, it would propagate the results from A twice to its
caller (the orchestrator itself).

By using a hashmap, we prevent this from happening on the JobTask level,
thus, artifacts are not getting reported to the user twice.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
Tested-by: Matthias Beyer <mail@beyermatthias.de>