Change implementation to use HashMap for errors
This patch changes the implementation to use an HashMap<Uuid, Error> for
error propagation.

The rationale behind this is the same as with the change to HashMap for
the artifacts: Errors are not getting propagated twice if they arrive at
a job from different child jobs.

This is technically not possible yet, because we propagate errors only
to one parent. Though, if the implementation changes one day (which it
could), this is one thing less we have to think about.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>