Remove "tree" from submit
This removes the "tree" column from the "submits" table.

This is because we do not store the build-tree in the database anymore.
We don't actually need this feature and we can always re-build the tree
from an old commit in the repository.
Thus, this is not required anymore.

Also, it is less easy to do as soon as the internal implementation
changes from a "tree" structure to a "DAG" structure.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>