Remove `Artifact` type almost entirely
This patch removes the `Artifact` type almost entirely.

The type parsed its path to extract name and version.
This information was _never_ needed, only resulting in the restriction that the
path must be parsable (which technically is not required at all).

This patch boils down the `Artifact` type to the absolute minimum, as a baseline
for its complete removal.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>