Remove `Artifact` type
This patch follows-up on the shrinking of the `Artifact` type and removes it
entirely.

The type is not needed. Only the `ArtifactPath` type is needed, which is a thin
wrapper around `PathBuf`, ensuring that the path is relative to the store root.

The `Artifact` type used `pom` to parse the name and version of the package from
the `ArtifactPath` object it contained, which resulted in the restriction that
the path must always be

    <name>-<version>...

Which should not be a requirement and actually caused issues with a package
named "foo-bar" (as an example).

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>