Fix: Query must root in the packages table
This fixes the query, because it must start from the packages table and filter
for package namd and version, and then join the other tables.

The dependency filtering was removed because it was too complex at this points.
Also, the release finding was changed so that the Release object is retrieved
later in the process of artifact selection.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>