Fix: Only print direct dependencies
Because the JobTask object gets _all_ transitive dependencies as well, this
number might be higher than the expected (direct) dependencies, resulting in
something like

    Waiting (11/3)

which is not nice UI-wise. Hence, limit to direct dependencies here.
We cannot count transitive dependencies at this point, so this is the best way
to do it right now.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>