Fix: Return script errors properly
This patch fixes error returning from the JobHandle::run() implementation.

Somehow, in the many rewrites, the error returning resulted in code that did not
differentiate between script run errors and scheduling errors.

This patch fixes this by making the JobHandle::run() method return

    Result<Result<Vec<ArtifactPath>>>

whereas the outer error is the scheduling result and the inner error is the
script result.

The calling code was adapted accordingly.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>