Rewrite artifact searching in replacement algorithm
This patch rewrites the replacement searching algorithm, to try the staging
store first and then the release store.

It does so by sorting the artifacts by whether they are in the staging store or
not (hence the FullArtifactPath::is_in_staging_store() function).

It filters out not-found artifacts and returns only ones that were found in
either store.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>